+++ lib/shad.dart
@@ -203,6 +203,7 @@ export 'src/utils/position.dart';
 export 'src/utils/provider.dart' hide ProviderReadExt, ProviderWatchExt;
 export 'src/utils/provider_index.dart';
 export 'src/utils/responsive.dart';
+export 'src/utils/shadow.dart';
 export 'src/utils/states_controller.dart';
 export 'src/utils/text_editing_controller.dart';